extracted ValidatorSelector component
