+++ packages/playground/src/components/index.ts
@@ -2,3 +2,4 @@ export { CopyLink } from './CopyLink';
 export { GeoPosition } from './GeoPosition';
 export { ThemeSelector } from './ThemeSelector';
 export { Selector } from './Selector';
+export { ValidatorSelector } from './ValidatorSelector';
